casting

+++ Exams-42-Piscine/Level_02/union/union.c
@@ -10,19 +10,19 @@ int main(int ac, char **av)
     {
         while (av[1][i])
         {
-            if (hash[av[1][i]] == 0)
+            if (hash[(int)av[1][i]] == 0)
             {
                 write(1, &av[1][i], 1);
-                hash[av[1][i]] = 1;
+                hash[(int)av[1][i]] = 1;
             }
             i++;
         }
         while (av[2][j])
         {
-            if (hash[av[2][j]] == 0)
+            if (hash[(int)av[2][j]] == 0)
             {
                 write(1, &av[2][j], 1);
-                hash[av[2][j]] = 1;
+                hash[(int)av[2][j]] = 1;
             }
             j++;
         }
